requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
future==0.18.3
h11==0.14.0
httptools==0.5.0
idna==3.4
peewee==3.16.2
pefile==2021.9.3
pewee==0.1.0.dev9
psycopg2==2.9.6
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
